+++ packages/http-api/package.json
@@ -15,10 +15,13 @@
     "@hapi/boom": "^10.0.1",
     "@hapi/h2o2": "^10.0.4",
     "@hapi/hapi": "^21.3.2",
+    "@opencrvs/toolkit": "^1.9.5-rc.141fdf1",
+    "@types/jsonwebtoken": "^9.0.10",
     "@types/lodash": "^4.14.197",
     "@types/node": "^20.5.3",
     "hapi-pino": "^12.1.0",
     "jose": "^5.0.0",
+    "jsonwebtoken": "^9.0.3",
     "lodash": "^4.17.21",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2",
