+++ packages/http-api/src/async-search/async-search-handler.ts
@@ -5,11 +5,10 @@ import {
   maybeEncryptedAsyncSearchRequestSchema
 } from '../validations'
 import { fromZodError } from 'zod-validation-error'
-import { ValidationError } from '../error'
+import { AuthorizationError, ValidationError } from '../error'
 import { search } from '../sync-search/sync-search-handler'
-import { validateToken, AuthorizationError } from 'opencrvs-api'
 import { registrySyncSearchBuilder } from 'dci-opencrvs-bridge'
-import { parseToken } from '../auth'
+import { parseToken, validateToken } from '../auth'
 import { randomUUID } from 'node:crypto'
 import { type ReqResWithAuthorization } from '../server'
 import { withSignature } from '../crypto/sign'
