+++ packages/http-api/src/server.ts
@@ -2,9 +2,13 @@ import * as Hapi from '@hapi/hapi'
 import { HOST, PORT, DEFAULT_TIMEOUT_MS, NODE_ENV } from './constants'
 import { routes } from './routes'
 import pino from 'hapi-pino'
-import { ValidationError, error } from './error'
+import {
+  ValidationError,
+  DailyQuotaExceededError,
+  AuthorizationError,
+  error
+} from './error'
 import H2o2 from '@hapi/h2o2'
-import { AuthorizationError, DailyQuotaExceededError } from 'opencrvs-api'
 
 export async function createServer() {
   const server = new Hapi.Server({
